feat(shell): add a helper for git backups
Problem: I'm prone to forgetting to commit and/or push changes to the
various repos I'm using around the place. Also prone to forgetting
the repo exists at all (have I missed any??)

Solution: create a helper to do the checking for me

Note: automate the checker ... should it run on login?